Remove deprecated code Move duplicated functions to a common file Formatting
